requirements.txt
asgiref==3.5.2
dj-database-url==1.0.0
Django==2.2.5
django-widget-tweaks==1.4.12
Pillow==9.2.0
python-decouple==3.1
pytz==2022.2.1
sqlparse==0.4.2
wrapt==1.11.1
